Move action to System Controller Display template instead of throwing exception (task #3086)
