+++ src/Template/System/home.ctp
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Copyright (c) Qobo Ltd. (https://www.qobo.biz)
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright     Copyright (c) Qobo Ltd. (https://www.qobo.biz)
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
+ */
+?>
+<div class="container-fluid">
+    <div class="row">
+        <div class="col-xs-12">
+            <div class="jumbotron">
+                <h1><?= __('Homepage') ?></h1>
+                <p><?= __('Access control has not been defined. Please contact the system administrator.') ?></p>
+            </div>
+        </div>
+    </div>
+</div>
